Move wl::{display, error}, wl_registry::{global, bind} from main.rs to wl_registry.rs
